- Fix compiler warning - Better warnings

+++ sql/multi_range_read.cc
@@ -1122,7 +1122,7 @@ bool DsMrr_impl::setup_buffer_sharing(uint key_size_in_keybuf,
      (ulonglong2double(rowids_size) + key_buff_elem_size));
 
   size_t bytes_for_rowids= 
-    round(fraction_for_rowids * (full_buf_end - full_buf));
+    (size_t)round(fraction_for_rowids * (full_buf_end - full_buf));
   
   long bytes_for_keys= (full_buf_end - full_buf) - bytes_for_rowids;
